+++ Expanded Cheat Menu/Code/Misc/InfoPaneCheats.lua
@@ -710,7 +710,7 @@ function Dome:CheatCrimeEvent()
 		text = "CheckCrimeEvents",
 		value = Dome.CheckCrimeEvents,
 	}}
-	local c = 1
+	local c = #item_list
 	local Dome = Dome
 	for key,value in pairs(Dome) do
 		if type(value) == "function" and key:sub(1,12) == "CrimeEvents_" then